requirements.txt
asgiref==3.6.0
chardet==5.1.0
Django==4.1.6
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
